Fix file names. Relocate test utils. Added godoc comments. Update test to use testify/assert
