+++ gateway/server_test.go
@@ -8,13 +8,14 @@ import (
 	"testing"
 	"time"
 
-	"github.com/TykTechnologies/tyk/config"
-	il "github.com/TykTechnologies/tyk/internal/log"
-	"github.com/TykTechnologies/tyk/internal/netutil"
-	"github.com/TykTechnologies/tyk/internal/otel"
-	"github.com/TykTechnologies/tyk/user"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
+
+	"github.com/TykTechnologies/tyk/config"
+	"github.com/TykTechnologies/tyk/internal/netutil"
+	"github.com/TykTechnologies/tyk/internal/otel"
+	"github.com/TykTechnologies/tyk/test"
+	"github.com/TykTechnologies/tyk/user"
 )
 
 func TestGateway_afterConfSetup(t *testing.T) {
@@ -222,13 +223,13 @@ func TestGateway_SyncResourcesWithReload(t *testing.T) {
 }
 
 func TestGateway_getHostDetails(t *testing.T) {
-	type checkFn func(*testing.T, *il.BufferedLogger, *Gateway)
+	type checkFn func(*testing.T, *test.BufferedLogger, *Gateway)
 
 	var (
 		orig_readPIDFromFile = readPIDFromFile
 		orig_mainLog         = mainLog
 		orig_getIpAddress    = netutil.GetIpAddress
-		bl                   = il.NewBufferingLogger()
+		bl                   = test.NewBufferingLogger()
 		check                = func(fns ...checkFn) []checkFn { return fns }
 
 		// matches ipv6 and ipv4
@@ -237,7 +238,7 @@ func TestGateway_getHostDetails(t *testing.T) {
 		ipAddrPattern = `^((([a-f0-9]{1,4}):){7}([a-f0-9]{1,4})|(([a-f0-9]{1,4})(:([a-f0-9]{1,4})){0,6})?::(([a-f0-9]{1,4})(:([a-f0-9]{1,4})){0,6})?|((([a-f0-9]{1,4}):){5}[a-f0-9]{1,4}|(([a-f0-9]{1,4}):){0,5}:([a-f0-9]{1,4}))|(([a-f0-9]{1,4}:){0,6}[a-f0-9]{1,4}|(([a-f0-9]{1,4}:){0,6}:([a-f0-9]{1,4})){0,1})|([a-f0-9]{1,4}:){0,7}:|([a-f0-9]{1,4}:){0,6}[a-f0-9]{1,4}|(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?))$`
 
 		hasErr = func(wantErr bool, errorText string) checkFn {
-			return func(t *testing.T, bl *il.BufferedLogger, gw *Gateway) {
+			return func(t *testing.T, bl *test.BufferedLogger, gw *Gateway) {
 				logs := bl.GetLogs(logrus.ErrorLevel)
 				if !wantErr && assert.Empty(t, logs) {
 					return
@@ -256,7 +257,7 @@ func TestGateway_getHostDetails(t *testing.T) {
 		}
 
 		hasAddress = func(addr string) checkFn {
-			return func(t *testing.T, bl *il.BufferedLogger, gw *Gateway) {
+			return func(t *testing.T, bl *test.BufferedLogger, gw *Gateway) {
 				matched, err := regexp.MatchString(addr, gw.hostDetails.Address)
 				if err != nil {
 					t.Errorf("Failed to compile regex pattern: %v", err)
