+++ test/buffered_logger.go
@@ -1,4 +1,4 @@
-package gateway
+package test
 
 import (
 	"sync"
@@ -7,19 +7,23 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// implements a buffered logger mainly to be used in tests
+// Implements a buffered logger to be used in tests in replacement of the main logger.
+// It stores the logs in memory and allows to retrieve them.
 
+// BufferedLog is the struct that holds the log entry.
 type BufferedLog struct {
 	Message string       `json:"message"`
 	Time    time.Time    `json:"time"`
 	Level   logrus.Level `json:"level"`
 }
 
+// BufferingFormatter is the struct that holds the buffer of the logs.
 type BufferingFormatter struct {
 	bufferMutex sync.Mutex
 	buffer      []*BufferedLog
 }
 
+// Format is the method that stores the log entry in the buffer.
 func (f *BufferingFormatter) Format(entry *logrus.Entry) ([]byte, error) {
 	f.bufferMutex.Lock()
 	defer f.bufferMutex.Unlock()
@@ -34,11 +38,13 @@ func (f *BufferingFormatter) Format(entry *logrus.Entry) ([]byte, error) {
 	return nil, nil
 }
 
+// BufferedLogger is the struct that holds the logger and the buffer.
 type BufferedLogger struct {
 	*logrus.Logger
 	bufferingFormatter *BufferingFormatter
 }
 
+// GetLogs returns the logs that are stored in the buffer.
 func (bl *BufferedLogger) GetLogs(Level logrus.Level) []*BufferedLog {
 	bl.bufferingFormatter.bufferMutex.Lock()
 	defer bl.bufferingFormatter.bufferMutex.Unlock()
@@ -54,6 +60,7 @@ func (bl *BufferedLogger) GetLogs(Level logrus.Level) []*BufferedLog {
 
 var buflogger *BufferedLogger
 
+// NewBufferingLogger creates a new buffered logger.
 func NewBufferingLogger() *BufferedLogger {
 	if buflogger != nil {
 		return buflogger
@@ -70,6 +77,8 @@ func NewBufferingLogger() *BufferedLogger {
 	return buflogger
 }
 
+// ClearLogs clears the logs from the buffer.
+// IMPORTANT: Must be called before each test iteration.
 func (bl *BufferedLogger) ClearLogs() {
 	bl.bufferingFormatter.bufferMutex.Lock()
 	defer bl.bufferingFormatter.bufferMutex.Unlock()
